AbstractPanel efficiency improvements
AbstractPanel has two new tweaks that can be used to speed up the
loading of plots:
	1) boolean beEfficient - checks to see if the same dot/size/color combo
has already been drawn and doesn't repeat it (default is true); this
leads to a dramatic improvement in QQPlots where millions of datapoints
can be reduced to a couple thousand; this does not remove the points, so
they are still included in the lookup, if that is active
	2) boolean createLookup - determines whether the lookup of nearest
neightbors (used with mouse click) is created and is now editable from
the constructor of the subclass of AbstractPanel; the default is to
include, but I have set it to false in the QQPanel constructor until we
actually use it to highlight/copy the marker name (which isn't currently
stored)
	
We still have the problem with an initial double rendering for all
AbstractPanels (but not when zooming). createImage() is being called
twice, even though it should only be called once. Forcing the second
call not to happen leads to a blank screen.

AbstractPanel has a DEBUGGING constant that can be turned on for more
logging to System.out

AbstractPanel has moved from a boolean finalImage to an integer that
allows for the intermediate stage to be recognized (though this hasn't
proved useful at all yet)

setFinalImage() affects ForestPanel, MosaicPanel, and SexPanel

QQPlot now requires a Logger