Refactor CLI to instance-based
Since CLI use required keeping references to an Objects instance and a
Map instance, which always correlated 1:1 with a CLI parsing workflow,
it makes more sense to use a single wrapper instance.

Using a CLI instance instead of static methods also allows simplified
method calls and is generally just easier on the eyes.