Refactor RegionNavigator
The RegionNavigator class has been modified to better centralize
navigation. It is intended to minimize repetition and complexity in
event listening between classes, allowing for standardized UI components
to be added with as little extra effort as possible (focusing purely on
navigational code).

As CompPlot did not actually have "Region" navigation built in yet, and
was lacking generic chromosome navigation, this first phase of
refactoring focuses on chromosome navigation - which is now unified
between CompPlot and Trailer.