Look up DNA id when matching CNVRectangles
When building CNVRectangles in CompPlot we use the IID and potentially
FID + IID to check sample compatibility. This was working by chance when
the DNA = the IID, but causing all rectangles to be hidden when they
were different.

Using the SampleData we can go from IID/FID to DNA and verify when a
region is part of a requested sample list.

Closes #106