Use explicit region lists when opening Trailer
The code path for opening the CNV lists as region and setting the first
sample and position seems to be broken. However, it is less robust
anyway than just using the newer explicit region list constructor - so
this is now what we do, e.g. opening trailer from sex plot.