Start factoring out common CLI arg/descs
This adds some consistency between main methods. Also, in general, since
the arg value has to be used twice (once to set the arg and once to
retrieve the value) it is safest to use a constant, and annoying to
redefine that constant in each class that accepts a project.