Make required header public visibility
This header is used in the MergeDatasets class. In general, having
mutable "constants"  like arrays with public visibility is dangerous (as
the contents of the array can be modified even though the reference is
final). One way to get around this is with a method that re-creates the
array with every call but that is potentially expensive. For now, I just
restored public visibility.