Fix MAD calc and display
TreeList is NOT a sorted tree, as I mistakenly believed. So the median
calc was not correct. I shifted back to a map of values to counts and
implemented a more intelligent median calculation to avoid the
computatino problems I initially ran into when converting the values to
lists for each calculation.

Also adjusted the Y axis buffer since 0 values were being cut off.