Refactor versioning
Update Genvisis to use a semver library for version parsing. Version
checking/display has been updated to ignore the SNAPSHOT portion of a
check during startup (and in the main window display).

With this change we can move our main version number (in Assembly right
now) to SNAPSHOT versioning without seeing this (often confusing for
users) term in the UI.

A more thorough version check should be part of the updater.