Temporary solution to reliance on dangerous old MarkerSet standard
This change partially reintroduces an issue with MarkerSet where the
internal array for names, chrs, and positions were returned. This will
act as a temporary solution to places where this behavior was relied on
(e.g. calling the getter for these arrays in every iteration of an array
rather than setting a local)

With this change, the internal data is not exposed but a SoftReference
to each array is held meaning it is likely that the same array will be
returned multiple times.