Add getDuplicateSets() to SampleData
Allows for access to the sets of duplicates stored in the Sample Data by
SampleQC Annotation

This change also encompasses some hardening of the SampleData class.
Most of the fields are now private and when the constructor fails an
exception is thrown instead of relying on the failedToLoad field. The
three places that used this field now catch the exception for their
error case.