Remove clashing lines from README
These lines are not properly formatted markdown. When viewing the README
they create a single paragraph with redundant links that go nowhere
helpful, with repeated mouse-over text.

It seemed like this text was the start of a "highlights" section, so it
was replaced with a link to the Genvisis manual.