Clean up selections when removing items
When removing a selected node or all nodes, the selections state array
needs to be updated in case any selected items were removed.

Unselecting and nulling out selected items fixes problems where TwoDPlot
tries to select items that do not exist.

Note that we can not change the size of the selections array because
that is what is enforcing the max number of selected items.