Add a Java version check
Note: this is tricky.

To have preserve runnability of a Java application in older JREs, the
entry point main class must be compatible with any desired legacy Java
versions.

Here we ensure this by modifying the GenvisisMain class to be compiled
with Java 6. The manifest for the uberjar retains the JVM version
specified in pom-genvisis, allowing all other components to be compiled
with the desired Java version.

This enables GenvisisMain to compare the runtime version to the manifest
version (as long as the runtime is at least compatible with the version
specified in the Assembly pom.xml).

Caveats:
* Any code paths in the GenvisisMain version check must be insulated
  from later Java versions.
* Genvisis has many entry points, and this does nothing to protect them.