Fix TwoDPlot accidental exclusion
Instead of lookup up samples solely on IID, we try DNA first and
foremost. Failing that, we can try FID\tIID using SampleData#lookup.

By only trying until we get a hit we avoid a bug which was always
iterating over fields and always trying IID (which was stored earlier in
the list of identifying data). If someone with that IID (but different
FID) had been excluded, most individuals would be excluded.

This is a partial fix for TwoDPlot, but see #281 for part of the
underlying problem (although SampleData#lookup was not being used
previously)