Create TaskManager framework
The goal of this framework is to create an agnostic mechanism for
performing computationally-intensive operations with arbitrary feedback
(whether via ProgressBar, ProgressMonitor, Loggers, etc...).

Such operations require special care when running in graphical
environments, to avoid blocking the EDT (causing a UI freeze) but still
allowing graphical updates (in the case of JProgressBars or text
output).

The SwingWorker + PropertyChangeListener paradigm has been established
to enable precisely this behavior: running intensive tasks off-EDT,
while providing hooks for progress and post-run actions.

The TaskManager framework facilitates proper use of of SwingWorkers and
PropertyChangeListeners without the need for the developer to maintain
references to both components at all times. In addition to simplifying
the logic required, this enables e.g. reporting to a master progress bar
without exposing a reference to that bar.

This framework consists of two core components:
* Tasks wrap SwingWorkers and are used as such. They provide an
  identifier to link with a listener.
* TaskListeners listen to events fired by Tasks and respond
  appropriately, e.g. by updating UI elements.

The relationship between Tasks and TaskListeners is many-to-many,
allowing coalesced reporting of multiple tasks to a single UI element,
and/or an individual Task reporting to multiple behaviors (e.g. both
logging and UI updates).

The intermediate between these two components is the static TaskManager
utility class, which allows Tasks and TaskListeners to be registered
without knowledge of each other. This encapsulation allows a separation
of UI and computation concerns.

The TaskTests class provides examples of use.

NB: while SwingWorkers were designed for GUI work, they do work
headlessly. This allows the framework to be used without regard for
whether an algorithm will run graphically or not.