Create benchmarks for indexFactors
Create tests for the indexFactors methods with 1D and 2D input arrays
using a search space large enough to demonstrate performance
implications of the current implementation.

These tests estalblish the expected results (hash of resulting index
array). Also, a naive map implementation is provided to compare
performance (which appears to be more than an order of magnitude
faster).