Fix default upload command
The dir command is not present on OSX by default, so this build was
breaking on mac. "echo" however is present on linux, windows and mac! So
this should work better.

This is definitely a hack - the upload and copy functions would be
better encapsulated in their own profiles. But I couldn't find a way to
do that and still have a nicely toggle-able genv profile (since there's
no way to do condition/anding of profiles)