Convert 2D indexFactors to use maps
As was done for the 1D indexFactors, using maps in the 2D method can
lead to significant performance improvements (see TestIndex class).
IndexFactors is used throughout Genvisis, including searches through
markers which was negatively affecting performance.

Unfortunately we can only use maps when an exact match is provided. This
could be revised with #330.