Fixes -parse issue when dummy files exist
parse was grabbing the header of the first file and assuming it 
for all other files when concatenating. This would cause alignment
issues if the first file was a dummy (had the default header) and
populated files down the line had additional columns.
In the process of fixing this, dummy files were eliminated, and
code that assumed the presence of a file for every chromosome was
removed.