Fix boundary checks in Compression.lrrCompress
This method was mistakenly setting the "high" float cutoff to a negative
value and the "low" cutoff to positive. It was also performing delta
checking incorrectly, taking the absolute value of only the lrr value,
instead of the absolute value of the different between the bounds.

This was causing all non-NaN lrr's (NaN's hit a different control
branch) to be identified as "outside of range". Since we are creating a
map entry with a unique key for each "outlier" in a project, we
effectively created a memory leak and made sample parsing unlikely to
succeed.

To avoid accidentally flipping the cutoffs again, the bound check is
simplified a single delta test (testing if the difference between two
absolute values is more than a delta of 0.0001f).