Adjust project creation gui sizing
The file selection buttons were being cut off, probably due to using
hard-coded dimensions. By instead packing first and making subsequent
slight adjustments (within the bounds of the screen) we get a window
that is aesthetically reasonable and will keep up as components are
added or removed, by relying on the layout manager.