Sanitizes ArrayUtils.toStr's nullValue to avoid NullPointerExceptions
The method will now join the string "null" in place of null values.
This won't change the output of anything currently calling the method
as it would've previously crashed with a NullPointerException in Joiner
but this may not be the ultimate desired behavior (other options
include skipping null values if there is no replacement value).