Staticify FishersExact2by2Calculator
It is not meaningful to track table counts between individual p-value
computations, so converting this functionality to a static utility class
makes it a bit more concicse (removing the need to instantiate in each
place of use, and always perform a load call followed by a compute
call).

The state that is helpful is the table of cached log factorial values.
Converting this to a list allows the ability to keep a single reference
which can be dynamically expanded as needed, and reused between
computations.