Create utility for consistent resource loading
There are a few ways to try and load bundled resources but the safest is to use
Class.getClassLoader.loadResource - as this works with unit tests and child class
loaders.

The Bundled utility methods allow consistent resource loading with
class.getClassLoader.getResource. It provides the convenience of not having to
specify a class to use as a loading reference, while allowing configuration of the
base class if needed.

These methods also allow consistent creation of Files and InputStreams. In particular,
running the resource URL through a URI before creating a file fixes loading issues
on Windows where the file path has unusual characters (e.g. spaces, %20).

Relevant use cases have been updated