Don't modify file name in icon loading
Adding a leading '/' is unexpected, and actually was preventing proper loading of
resources due to the file no longer being found.

Updated the logic to try loading the filename from resources directly and if not
found, see if the file exists and load it.

Typically we'd expect icons to be loaded from bundled resources.