Switch auto copy and upload to conditional statements
Trying to reference an application that always exists and copying in place is costly,
fragile and dumb. Instead we can use the ant-contrib if syntax to test if these
properties actually exist and respond appropriately.

Updated the corresponding section of the README which was out of date. It is now more
general, allowing for specific examples for Eclipse to be written in a wiki.