Fix tar.gz resource checking
The local check for a TarGzResource needed to check if the archive had been
downloaded in the isLocallyAvailable check. Failure to do so was causing
incorrect checks of the md5 and the download success. These resources now report
themselves as locally available after downloading the archive, and the get method
unwraps them as intended.

When downloading a resource we now properly return false if we can't compute
the local md5 after the download.

Also exposed the log and name methods for convenience.

Closes #441