AnnotateSampleDataStep: fix output check
The Exclude and Duplicate headers were reversed in where they should be checked.

Also, if there is no duplicate file produced by plink, the associated headers will
not be added, so we shouldn't check for them in the outputExists method.

This fix should enable AnnotateSampleDataStep to be used from the GUI, and properly
reflect its completion status.