Combine implementation of filterBasedOnNumberOfCNVsAtLocus[InMemory]
Removes unneeded cnvsAsPositions flag. With CNV endpoints not required
to be markers, this path is no longer necessary.

The non-InMemory flavor does now hold all CNVs in memory prior to
printing. If this proves to be a problem, a more intricate splitting of
functionality can be made but this was the most straightforward way to
reduce the use of copied code.