Convert Mosaicism to expunge reliance on chr-pos order of markers
Uses MarkerDetailSet methods to remove parallel array based processing
as much as possible.

Includes a quick fix to use a Set<Marker> for mosaicism from Sex Checks,
this will be done properly when SexChecks is also converted to
MarkerDetailSet and made to not rely on chr-pos ordering.

A small bug is resolved by these changes where chromosomes with no
markers in the p arm were skipping the first marker for the chromosome
when indexing the q arm. Because of this, results for these chromosomes
will vary slightly after this change but are technically correct while
the old results were technically incorrect.