Make WorkerTrain hasNext safer
hasNext relies on checking the queue of tasks + the set of producers. As long as
one of these still has something to do, hasNext is true.

However, there is a potential race condition where an item could be removed from
the producers but not yet added to the queue, in which case hasNext could return
false, ending iteration even though there are unprocessed elements.

Sychronizing the transfer to the queue with the hasNext check resolves this issue.

Closes #474