Add InterruptingExecutorService and RunService utilities
On a call to execute(), InterruptingExcutorService cancels the current
runnable and submits the new runnable.  Methods called within submitted
Runnable objects must be instrumented with InterruptedExceptions and
try-catch blocks for the same to work properly.

RunService wraps an InterruptingExecutorService and a single final
Runnable; when the run() method is called, this same runnable is
re-submitted to the execute() command.